+++ src/EVM/Fetch.hs
@@ -263,10 +263,7 @@ getSolutions solvers symExprPreSimp numBytes pathconditions = do
       collectSolutions :: Expr EWord -> Prop -> Config -> IO (Maybe [W256])
       collectSolutions symExpr conds conf = do
         let smt2 = assertProps conf [(PEq (Var "multiQueryVar") symExpr) .&& conds]
-        let maxNum = case conf.maxExplore of
-                        Just x -> Prelude.min x conf.maxBranch
-                        Nothing -> conf.maxBranch
-        checkMulti solvers smt2 $ MultiSol { maxSols = maxNum , numBytes = numBytes , var = "multiQueryVar" }
+        checkMulti solvers smt2 $ MultiSol { maxSols = conf.maxWidth , numBytes = numBytes , var = "multiQueryVar" }
 
 -- | Checks which branches are satisfiable, checking the pathconditions for consistency
 -- if the third argument is true.